# Pipe images to ffmpeg to speed up a lot