--- a/requirements.txt
+++ b/requirements.txt
@@ -1,4 +1,5 @@
 Pillow==6.2.0
+pexpect==4.8.0
 
 # python-midi 0.2.2 is also required; however, the upstream repository
 # still has an unmerged pull request fixing a significant tempo