After profiling using cProfile, it seems that most time is used to compress images into PNG format.

Orignal time:

```
87.53s user 2.42s system 125% cpu 1:11.65 total
```

Use BMP format instead of PNG can reduce most time:

```
24.09s user 5.17s system 145% cpu 20.075 total
```

Using BMP format can reduce CPU time, but it use much more disk space (22MBps per image file).  So I use ``ffmpeg -f image2pipe`` insted of ``ffmpeg -f image2`` to pipe images to ffmpeg, avoid writing to disk.  Because the ``subprocess`` module doesn't support multiple write to stdin, so I introduce ``pexpect`` module to pipe multiple images one by one to ffmpeg.  After that, the CPU time is:

```
25.96s user 3.93s system 192% cpu 15.529 total
```

After removing unnecessary ``copy()``, the total time is reduced one more second:

```
24.97s user 3.53s system 200% cpu 14.212 total
```

I have test it on MacOS for python2 and python3.